requirements.txt
cryptography==42.0.7
Pillow==10.3.0
PyQt5==5.15.10